j'ai termine l'exercice 8

+++ UniversiteDomain/Dtos/NoteAvecUeDto.cs
@@ -7,7 +7,7 @@ public class NoteAvecUeDto
     public long EtudiantId { get; set; }
     public long UeId { get; set; }
     public UeDto UeDto{get; set;}
-    public float Valeur { get; set; }
+    public double Valeur { get; set; }
 
     public NoteAvecUeDto ToDto(Note note)
     {
